Removed axios & sass, minor improvements to provenance function
