+++ packages/prez-ui/app/utils/types.ts
@@ -1,0 +1,15 @@
+export type SPARQLResultsJSON = {
+    head: {
+        vars?: string[];
+        link?: string[];
+    },
+    results?: {
+        bindings: Record<string, {
+            type: "uri" | "literal" | "bnode";
+            value: string;
+            "xml:lang"?: string;
+            datatype?: string;
+        }>[];
+    },
+    boolean?: boolean;
+};
